+++ apps/example-vite-project/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @notjustcoders/example-vite-project
 
+## 0.0.3
+
+### Patch Changes
+
+- Fix: WrappedClass was not binding the this context of the original class
+- Updated dependencies
+  - @notjustcoders/one-logger-client-sdk@1.0.3
+
 ## 0.0.2
 
 ### Patch Changes