+++ packages/logger/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @notjustcoders/one-logger-client-sdk
 
+## 1.0.3
+
+### Patch Changes
+
+- Fix: WrappedClass was not binding the this context of the original class
+
 ## 1.0.2
 
 ### Patch Changes